feat(build): add gitlab-ci job for building of the oci image
Notes: The current implementation of this build job is very simple.
It simply starts a new build process from scratch, possibly utilizing the
pre-built artifacts from cache.nixos.org, and exports the resulting OCI image as
an artifact.

This can be improved in multiple ways, e.g. by utilizing the persistent caching
mechanism of Nix through a persistent building container (see build.sh for a
local implementation of this).